chore(variadic_from): Clean up test directory and refactor lib structure
