+++ module/core/variadic_from/tests/inc/mod.rs
@@ -1,13 +1,2 @@
-#![ allow( unused_imports ) ]
-
 use super::*;
-
-// mod derive_test; // Commented out for incremental re-enabling
-
-// #[ cfg( all( feature = "derive_variadic_from", feature = "type_variadic_from" ) ) ]
-// #[ path = "./compile_fail/err_from_0_fields.rs" ]
-// mod err_from_0_fields; // Commented out for incremental re-enabling
-
-// #[ cfg( all( feature = "derive_variadic_from", feature = "type_variadic_from" ) ) ]
-// #[ path = "./compile_fail/err_from_4_fields.rs" ]
-// mod err_from_4_fields; // Commented out for incremental re-enabling
+use test_tools::exposed::*;
